Cast to long long before left-shifting too much.
When left-shifting we must be sure that the value being
shifted is large enough to not lose bits.
The 'chunkssize' in CreateBitmap is only 'long' so it
can overflow.  So cast to 'long long' first.

Also fix a similar issue in Detail even though it isn't currently
being compiled.

Reported-by: Tomasz Chmielewski <mangoo@wpkg.org>

+++ Detail.c
@@ -409,7 +409,7 @@ int Detail(char *dev, int brief, int export, int test, char *homehost)
 #if 0
 This is pretty boring
 			printf("  Reshape pos'n : %llu%s\n", (unsigned long long) info.reshape_progress<<9,
-			       human_size(info.reshape_progress<<9));
+			       human_size((unsigned long long)info.reshape_progress<<9));
 #endif
 			if (info.delta_disks > 0)
 				printf("  Delta Devices : %d, (%d->%d)\n",
